Inserted biolerplate to start writing tests
I will be writing tests using the asunit framework with my own library built on top
Going forward i plan to write tests and production code concurrently following a TDD
style. The test writen for code already in the system will be weaker because
they are tacked on after the fact, i will need to be cautious to not let any errors
slip by.